fix path overlap judgement
