+++ node-commons/src/main/java/org/apache/iotdb/commons/path/PartialPath.java
@@ -307,6 +307,31 @@ public class PartialPath extends Path implements Comparable<Path>, Cloneable {
     return true;
   }
 
+  /**
+   * Test if this path pattern overlaps with input path pattern. e.g. "root.sg.**" overlaps with
+   * "root.**", "root.*.d.s" overlaps with "root.sg.d.s", "root.sg.**" overlaps with "root.**.s",
+   * "root.*.d.s" doesn't overlap with "root.sg.d1.*"
+   *
+   * @param rPath a plain full path of a timeseries
+   * @return true if a successful match, otherwise return false
+   */
+  public boolean overlapWith(PartialPath rPath) {
+    String[] rNodes = rPath.getNodes();
+    for (int i = 0; i < this.nodes.length && i < rNodes.length; i++) {
+      if (nodes[i].equals(MULTI_LEVEL_PATH_WILDCARD)
+          || rNodes[i].equals(MULTI_LEVEL_PATH_WILDCARD)) {
+        return true;
+      }
+      if (nodes[i].equals(ONE_LEVEL_PATH_WILDCARD) || rNodes[i].equals(ONE_LEVEL_PATH_WILDCARD)) {
+        continue;
+      }
+      if (!nodes[i].equals(rNodes[i])) {
+        return false;
+      }
+    }
+    return true;
+  }
+
   @Override
   public String getFullPath() {
     if (fullPath == null) {
