+++ server/src/main/java/org/apache/iotdb/db/mpp/common/schematree/PathPatternTree.java
@@ -259,7 +259,7 @@ public class PathPatternTree {
     PartialPath pattern = prefixPath.concatNode(MULTI_LEVEL_PATH_WILDCARD);
     List<PartialPath> involvedPath = new ArrayList<>();
     for (PartialPath path : pathList) {
-      if (pattern.matchFullPath(path) || path.matchFullPath(pattern)) {
+      if (pattern.overlapWith(path)) {
         involvedPath.add(path);
       }
     }
